+++ simulators/c++2/src_simulator/app/TMLEventFBChannel.h
@@ -85,9 +85,9 @@ public:
 
 	void write(TMLTransaction* iTrans){
 #ifdef LOSS_ENABLED
-		if (this->_maxNbOfLosses > this->_nbOfLosses && this->_lossRate!=0 && myrand(0,99) < this->_lossRate){
+	  if (this->_maxNbOfLosses > this->_nbOfLosses && this->_lossRate!=0 && (unsigned int)myrand(0,99) < this->_lossRate){
 			this->_nbOfLosses++;
-		}else{
+		} else {
 #endif
 			this->_content++;
 			if (paramNo!=0){
